+++ src/frontend/src/features/chat/components/Entry.tsx
