+++ src/frontend/src/features/rooms/livekit/components/SidePanel.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next'
 import { ParticipantsList } from './controls/Participants/ParticipantsList'
 import { PanelId, useSidePanel } from '../hooks/useSidePanel'
 import React, { ReactNode, useCallback, useRef } from 'react'
-import { Chat } from '../prefabs/Chat'
+import { Chat } from '@/features/chat/components/Chat'
 import { Effects } from './effects/Effects'
 import { Admin } from './Admin'
 import { Tools } from './Tools'
